Tweak patch to build using just configure
HACK: The Debian packaging just patches lib/Makefile.inc and that's
enough to add a new file (fopen.c) in Debian packaging, which will do
a lot more autofoo magic.

But we want this new file to also build in another context where we
just use configure. Extend the patch file to add the extra bits into
lib/Makefile.in that we need for fopen.c to work.